Refactor: `interface` 분리
