+++ src/client/src/api/auth.ts
@@ -1,15 +1,6 @@
+import { SignInAPIProps, SignUpAPIProps } from "../interfaces/interface";
 import publicApi from "./public_api";
 
-export interface SignUpAPIProps {
-    email: string;
-    password: string;
-    company: string;
-}
-
-export interface SignInAPIProps {
-    email: string;
-    password: string;
-}
 
 export const signup = async ({email, password, company}: SignUpAPIProps) => {
     const response = await publicApi.post('/api/v1/auth/signup', {
